TryPares not all

+++ Calculator/Method.cs
@@ -7,65 +7,176 @@ namespace Calculator
 {
     class Method
     {
-        
-        public void BeginPL(double numberfirst, out double result, out double numbersecond)
-        {           
-            double.TryParse(Console.ReadLine(), out numbersecond);
-            result = numberfirst + numbersecond;
-        }
-        public void BeginSB(double numberfirst, out double result, out double numbersecond)
+        public string input;
+        public void BeginPL(char sign, double numberfirst, out double result,out double numbersecond)
         {
-            double.TryParse(Console.ReadLine(), out numbersecond);
-            result = numberfirst - numbersecond;
+            while (true)
+            {
+                Console.Clear();
+                Console.WriteLine($"{numberfirst} {sign} number 2");
+
+                input = Console.ReadLine();
+                bool rezultat = double.TryParse(input, out numbersecond);
+                if (rezultat)
+                {
+                    result = numberfirst + numbersecond;
+                    break;
+                }
+                else
+                {                    
+                    Console.Clear();
+                    Console.WriteLine("Eba ty lox BeginPL");
+                    Console.ReadKey();
+                }
+            }
         }
-        public void BeginMUL(double numberfirst, out double result, out double numbersecond)
+        public void BeginSB(char sign, double numberfirst, out double result, out double numbersecond)
         {
-            double.TryParse(Console.ReadLine(), out numbersecond);
-            result = numberfirst * numbersecond;
+            while (true)
+            {
+                Console.Clear();
+                Console.WriteLine($"{numberfirst} {sign} number 2");
+
+                input = Console.ReadLine();
+                bool rezultat = double.TryParse(input, out numbersecond);
+                if (rezultat)
+                {
+                    result = numberfirst - numbersecond;
+                    break;
+                }
+                else
+                {
+                    Console.Clear();
+                    Console.WriteLine("Eba ty lox BeginSB");
+                    Console.ReadKey();
+                }
+            }
+        }
+        public void BeginMUL(char sign, double numberfirst, out double result, out double numbersecond)
+        {
+            while (true)
+            {
+                Console.Clear();
+                Console.WriteLine($"{numberfirst} {sign} number 2");
+
+                input = Console.ReadLine();
+                bool rezultat = double.TryParse(input, out numbersecond);
+                if (rezultat)
+                {
+                    result = numberfirst * numbersecond;
+                    break;
+                }
+                else
+                {
+                    Console.Clear();
+                    Console.WriteLine("Eba ty lox BeginMUL");
+                    Console.ReadKey();
+                }
+            }
         }
         public void BeginDIV(char sign, double numberfirst, out double result, out double numbersecond)
         {
             while (true)
-            {
-                result = 0;
+            {                
                 Console.Clear();
                 Console.WriteLine($"{numberfirst} {sign} number 2");
-                double.TryParse(Console.ReadLine(), out numbersecond);
-                if (numbersecond == 0)
+                input = Console.ReadLine();
+                bool rezultat = double.TryParse(input, out numbersecond);
+                if(rezultat)
                 {
-                    Console.WriteLine("dadada");
-                    Console.WriteLine("continue ? esc = exit");
-                    if (Console.ReadKey().Key == ConsoleKey.Escape)
+                    if (numbersecond == 0)
                     {
-                        Process.GetCurrentProcess().Kill();
+                        Console.WriteLine("can not div on zero");
+                        Console.WriteLine("continue - press any key \nesc = exit");
+                        if (Console.ReadKey().Key == ConsoleKey.Escape)
+                        {
+                            Process.GetCurrentProcess().Kill();
+                        }
+                    }
+                    else
+                    {
+                        result = numberfirst / numbersecond;
+                        break;
                     }
                 }
                 else
                 {
-                    result = numberfirst / numbersecond;
-                    break;
+                    Console.Clear();
+                    Console.WriteLine("Eba ty lox BeginDIV");
+                    Console.ReadKey();
                 }
             }
         }
 
 
-        public void PL(ref double result, ref double rezultend, out double c)
-        {            
-            c = Convert.ToDouble(Console.ReadLine());
-            rezultend = result;
-            result += c;            
+        public void PL(char ex, ref double result, ref double rezultend, out double c)
+        {
+            while (true)
+            {
+                Console.Clear();
+                Console.WriteLine($"{result} {ex} number");
+                input = Console.ReadLine();
+                bool rezultat = double.TryParse(input, out c);
+                if (rezultat)
+                {
+                    rezultend = result;
+                    result += c;
+                    break;
+
+                }
+                else
+                {
+                    Console.Clear();
+                    Console.WriteLine("Eba ty lox PL");
+                    Console.ReadKey();
+                }
+            }
         }
-        public void SB(ref double result, ref double rezultend, out double c)
-        {            
-            c = Convert.ToDouble(Console.ReadLine());
-            rezultend = result;
-            result -= c;
+        public void SB(char ex, ref double result, ref double rezultend, out double c)
+        {
+            while (true)
+            {
+                Console.Clear();
+                Console.WriteLine($"{result} {ex} number");
+                input = Console.ReadLine();
+                bool rezultat = double.TryParse(input, out c);
+                if (rezultat)
+                {
+                    rezultend = result;
+                    result -= c;
+                    break;
+
+                }
+                else
+                {
+                    Console.Clear();
+                    Console.WriteLine("Eba ty lox PL");
+                    Console.ReadKey();
+                }
+            }                        
         }
-        public void MUL(ref double result, ref double rezultend, out double c)
-        {           
-            c = Convert.ToDouble(Console.ReadLine());
-            rezultend = result;
-            result *= c;            
+        public void MUL(char ex, ref double result, ref double rezultend, out double c)
+        {
+            while (true)
+            {
+                Console.Clear();
+                Console.WriteLine($"{result} {ex} number");
+                input = Console.ReadLine();
+                bool rezultat = double.TryParse(input, out c);
+                if (rezultat)
+                {
+                    rezultend = result;
+                    result *= c;
+                    break;
+
+                }
+                else
+                {
+                    Console.Clear();
+                    Console.WriteLine("Eba ty lox PL");
+                    Console.ReadKey();
+                }
+            }                                    
         }
         public void DIV(char ex, ref double result, ref double rezultend, out double c)
         {
@@ -73,24 +184,33 @@ namespace Calculator
             {
                 Console.Clear();
                 Console.WriteLine($"{result} {ex} number");
-                c = Convert.ToDouble(Console.ReadLine());
-
-                if (c == 0)
+                input = Console.ReadLine();
+                bool rezultat = double.TryParse(input, out c);
+                if (rezultat)
                 {
-                    Console.WriteLine("dadada");
-                    Console.WriteLine("continue ? esc = exit");
-                    if (Console.ReadKey().Key == ConsoleKey.Escape)
+                    if (c == 0)
                     {
-                        Process.GetCurrentProcess().Kill();
+                        Console.WriteLine("dadada");
+                        Console.WriteLine("continue ? esc = exit");
+                        if (Console.ReadKey().Key == ConsoleKey.Escape)
+                        {
+                            Process.GetCurrentProcess().Kill();
+                        }
                     }
+                    else
+                    {
+                        rezultend = result;
+                        result /= c;
+                        break;
+                    }                    
                 }
                 else
                 {
-                    rezultend = result;
-                    result /= c;
-                    break;
+                    Console.Clear();
+                    Console.WriteLine("Eba ty lox PL");
+                    Console.ReadKey();
                 }
-            }
+            }            
         }
